Change init_interp_attributes mask for mpasa60 to gx1v7 (instead of tx0.1v2)
The mpasa60 mask in the init_interp_attributes element was tx0.1v2. This was an
accident as the first mpasa60 grid was created incorrectly with that mask.
Instead, it should use the gx1v7 mask.

This commit simply corrects the init_interp_attributes element for the mpasa60
grid to use the gx1v7 mask.