Add mpasa480 fsurfdata, maps, and res to namelists
This commit adds references to the mapsa480 fsurfdata, fsurfdata maps, and adds
mpasa480 as a possible value to the res namelist definition. At present, this
commit does not contain a landuse.timeseries for the mpasa480 km grid.

The mpasa480 grid is a quasi-uniform, global mesh with 2,462 horizontal grid
columns (cells).