Add mpasa15-3-conus fsurfdata, landuse.timeseries and maps to namelist
This commits adds references to the mpasa15-3-conus fsurfdata,
landuse.timeseries, fsurfdata maps and adds mpasa15-3-conus as a possible value
to the res definition.

The mpasa15-3-conus is a 15 km to 3 km variable resolution, global grid with
circular refinement that is centered over CONUS. It has 6,488,066 horizontal
grid columns/cells.